+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - (20.08.2022)
+## Changes
+- added possiblity switch ports by port name
+- code cleanup
+- config schema update
+
 ## [0.11.1] - (19.08.2022)
 ## Changes
 - fix ssid hide by name